requirements.txt
aiohttp>=3.9.4
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-bootstrap-templates==1.1.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.2
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
idna>=3.7
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2>=3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
plotly==5.20.0
pycodestyle==2.11.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
retrying==1.3.4
scipy==1.13.0
six==1.16.0
tenacity==8.2.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug>=3.0.3
yarl==1.9.4
zipp==3.18.1
